+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [2.0.29] - 10.15.2020
 
-- Allow to configure new added script to be executed just on current fodler instead to run it across all dependencies
+- Allow to set configure new added script to be executed just on current fodler instead to run it across all dependencies
 - Add additional flag to turn off/on progress bar
 - Add additional flag specify the number of parallel jobs to use
+- Allow to set configure new added script to be executed in verbose mode